Minor changes to gain speed
With grid_point = 300, the program runs very slowly. So I change grid_point to 100 and L to 0.5 (this way dx is still .005). This makes the code runs a lot faster. However, the result was just a zoomed in version of the original data. Hence I created a constant called scale and multiply it in the real_space function. This fixes the problem and now the program also produces detail features that has not been seen before. Such as the bigger mass now has a tail develop after the "collision."